misc cleanup; compatability with linux gcc/clang & misc optimization (needs new unit tests); add ft_to_inverse (invert char case) and separate & rename static idigits to ft_nbr_len

+++ ft_lmax.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-int	ft_lmax(int a, int b)
+long	ft_lmax(long a, long b)
 {
 	if (a > b)
 		return (a);
